Update on "[quant][refactor tests] Move quantization tests into subfolders"
Summary:
Create folders for each test category and move the tests.
Will follow-up with a cleanup of test_quantization.py

Test Plan:
python test/test_quantization.py

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D28718742](https://our.internmc.facebook.com/intern/diff/D28718742)

[ghstack-poisoned]
